Fix bug: one player enter a weapon, both player picked up a gun.
Introduced SignalBus for global signals. Currently, pistol emit
the weapon_entered signal. And the player handles it.
This prevents the problem of having to connect to specific nodes,
when the node might not have been added to the scene tree yet.

NOTE: this also fixed the bug where the static pistol still exists
after a player picked it up. (But the process merits discussions)