Add automatic broadcast IP detection
The functionality is implemented in a c executable in the root
of this project. It returns the broadcast IP. The ServerBrowser
script executes this script to acquire the IP address.

Currently this function is only tested on macOS (UNIX like system).
I don't know what will happen if this is executed on windows yet.
But maybe if it doesn't work, I can call make in the script to
create the executable on the fly. But this is far from a good
approach.