Add a waiting room, allowing client to join and leave the room
Now when the host start hosting, or when the client joins the game,
they go to a waiting room. In the waiting room, the host has 1
button to start the game, and a button to cancel the host session.
The client has a button to leave the room. So only the host can
start a game.

Todo: All new buttons are implemented, except the "cancel host" button
on the server.