Fix everyone is a host in waiting room problem.
Before, the visibility of a waiting room is set when it enters the
scene. Since waiting room is part of the server_select screen, it
is set at first, which means everyone "seems" like a host.
Now the visibility of buttons are set again when "join" or "host"
is pressed.