Fix exit button. And waiting room icons
Before exit button clicks, but does not go back, or does not display
Search by IP button when return. This is now fixed.

During the fix, another bug was found: when entering and exiting
the waiting room multiple times, the previous character icons are
not cleared. This was not due to the GameManager not cleared.
Instead, it's because the waitroom is not reloaded but just hidden
so the previous contents are not cleared.
Now, when we refresh the players, we not onlly add new players,
but also remove old players that are not in the room no mo.