Add three game mode selection buttions, and a gotogame button
Only the host can choose the three buttons. But all players can
click on go to game button. (This can be disabled later)

all players can click on gotogame_btn, this is a UI compromise.
If I disable the button, the user cannot see if the host is
hovering/clicking on the button. They can only see that if their
own button is enabled.

We need to discuss what we want for the players to do.