Enable cancle host from all locations
Now server is no longer a child node of server select. It is a
global node. And there is now a new signal in signal bus for
cancel host. It can be called anywhere in the game to stop a
hosting session. When emitted, the global server node will disconnect
all players, and stop hosting.